requirements.txt
#Include needed libraries

google-generativeai